# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Pagamentos",
        "description": "API responsável por autenticação, saldo, depósitos, saques e transferências entre usarios comuns e lojiostas",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/transfer/reverse": {
            "post": {
                "tags": [
                    "Transferências"
                ],
                "summary": "Estornar transferência",
                "operationId": "7a7e4a7f66e3f878a5ea96228cc16891",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "transfer_id"
                                ],
                                "properties": {
                                    "transfer_id": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transferência estornada"
                    },
                    "422": {
                        "description": "Não foi possível estornar"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/transfer/send": {
            "post": {
                "tags": [
                    "Transferências"
                ],
                "summary": "Enviar transferência",
                "operationId": "07b50065f0eb0461cb5a7c4d57b28dc1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "receiver_id",
                                    "amount"
                                ],
                                "properties": {
                                    "receiver_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 75
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transferência realizada"
                    },
                    "422": {
                        "description": "Erro na transferência"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users": {
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Criar usuário",
                "operationId": "c585e630a62413459baeeb430678b499",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "cpf_cnpj",
                                    "type",
                                    "phone",
                                    "balance"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Felipe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "felipe@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123123123"
                                    },
                                    "cpf_cnpj": {
                                        "type": "string",
                                        "example": "12312312312"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "common",
                                            "merchant"
                                        ],
                                        "example": "common"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "5549999999999"
                                    },
                                    "balance": {
                                        "type": "number",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/v1/users/deposit": {
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Realizar depósito",
                "operationId": "a1afce0d005318b3e4fbc2be66a844a0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Depósito realizado"
                    },
                    "422": {
                        "description": "Valor inválido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login do usuário",
                "operationId": "e833ac3eab3be4e246ab7ced4e0db2a2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/api/v1/users/my-balance": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Consultar saldo do usuário",
                "operationId": "774ba02122208af8aa468692d3c625a1",
                "responses": {
                    "200": {
                        "description": "Saldo atual",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "balance": {
                                            "type": "number",
                                            "example": 1500.5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users/withdraw": {
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Realizar saque",
                "operationId": "79eb6a85d0b8e41d05ebbce6f197b08a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Saque realizado"
                    },
                    "422": {
                        "description": "Saldo insuficiente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Transferências",
            "description": "Transferências"
        },
        {
            "name": "Usuários",
            "description": "Usuários"
        },
        {
            "name": "Auth",
            "description": "Auth"
        }
    ]
}